backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pydantic==2.5.2
jinja2==3.1.6
httpx==0.27.0
beautifulsoup4==4.12.3
